Improvement of PyFunceble's bahavior.
Touches #57.
Touches Ultimate-Hosts-Blacklist/dev-center#42.

Before this patch, PyFunceble may returned some false positive.
In fact, if a subject didn't have and NS record, we didn't took in
consideration the CNAME and DNAME records which acts as alias of the
given subject. That caused some domains for being flagged as INACTIVE
when there are in reality ACTIVE.
From now, PyFunceble will try to get the NS records and failback to
CNAME or DNAME in case there was no NS record or we could not extract
it.

From now, PyFunceble will not systematically execute all logic(s).

Indeed, before this patch, PyFunceble was systematically running the
following (if authorized):
  * WHOIS // Expiration date extraction
  * DNS Lookup
  * HTTP Code extraction
  * Extra rules (based on previously collected data)

From now, PyFunceble do the following (if authorized):
  * WHOIS // Expiration date extraction
  * DNS Lookup
    * If WHOIS wasn't exploitable.
  * HTTP Code
    * If WHOIS and DNS Lookup did not succeed.
  * Extra rules (based on previously collected data)

From now, HTTP Code play an even more important role.

Indeed, before this patch, it was important but not completly used. From
now it's completly used. In fact, if WHOIS and DNS Lookup are not
exploitable, but the HTTP Code is still given, the status
will be automatically escalated to ACTIVE.
That was not the case as we were previously more focused on the
extraction of the NS record as fallback. This don't break anything. It's
just an extra safety check.

From now, we provide a more consistant `_status`, `status`,
`_status_source` and `status_source` information.
Indeed, with the restructuration branch which was merge last week, it
became inconsistent.

It is not important for the rest of the world, but, it is important for
the @ultimate-hosts-blacklist infrastructure as it provides a
`volatile.list` which sits on top of the consistancy of the previously
mentionned indexes.

Contributors:
  * @dnmTX
  * @spirillen