Improvement of the workflow behind our copy of the IANA and PSL database
Indeed, before this patch, I had to manually update the following files
with the help of `--iana` and `--public-suffix`:
  * `iana-domains-db.json`
  * `public-suffix.json`

From now, they will be automatically updated every 24 hours at the
following locations:

  * https://github.com/PyFunceble/iana
  * https://github.com/PyFunceble/public-suffix

This will hopefully make PyFunceble a bit more reactive to real-world
changes.

As reaction to the 2 new locations, the previously mentionned files are
deleted from this repository.

Also:
  * From now, the download of the previously mentionned files will be
    systematically done if they are not present.
  * From now, the download time will be savef into a new file called
    `.pyfunceble_intern_downtime.json`
  * From now, the download of the previously mentionned files will be
    re-downloaded if the previous downloaded time was in about 24 hours.

    Note:
      * This patch does not change any of the previous behavior.
        Indeed, if the previously mentionned files are deleted, they
        will be automatically redownloaded.