Some rewrite as it felt heavy
I have altered some of the text to easing the reading.

This line (43) I'm not all to sure of however:
`- Set the branch to which the CI will push the result to. For the autosaving system.`
`- Set the branch to push the result to, for the autosaving system.`

There can be many variation here. I'm all open for suggestions.