Introduction the .PyFunceble.overwrite.yaml file.
Before this patch, the configuration management may have been
hard to end-user. Therefore, I hereby introduce the
`.PyFunceble.overwrite.yaml` file.

When a file named after it is found into the given or pre-configured
configuration location, it will be read and loaded into PyFunceble.
In contrast to the already known and existing configuration file, this
file will never be written by PyFunceble.

End-user are responsible for its maintenance.